feat: Set root router.
Create a new file called 'Home.jsx' and render the simple string 'Home'.
Create a new file called 'Error.jsx' and render the notificaiton text.
Import the 'createBrowserRouter' from the 'react-router-dom' and pass an array of 'Route' object as an argument.
Add the 'errorElement' property to the array of 'Route' objects.
Add the nested 'children' property to the array of 'Route' objects.
Render the component based on the routes using the 'Outlet' component from 'react-router-dom' in the 'App' component.
In the first nested children of 'Routes' object, set the 'index' property to true and render the 'Home' component to handel the index router.