feat: Login with google.
Import the Firebase app instans from the firebase config file.
Initialize the auth and db service using the app instance.
Define the handleClickLogin function to handle the login with google.
Define the handleClickLogout function to handle the logout.
Fetch the 'admins' data from the 'realtimedatabase' to check if the logged-in user is an admin.
Pass the 'isLogin', 'isAdmin', 'userName', 'userProfile', 'handleClickLogin' and 'handleClickLogout' values as props to consuming component.
Export the useFirebase constom hook.