feat: Add product to cart.
When the user clicks the 'Add to Cart' button:

Check if the user is logged in. If not, display an alert message asking the user to login.
Confirm with the user that they want to add the product to the cart.
If the user confirms, check if the product with the same options already exists in the cart.
If the product exists, confirm with the user if they want to add another instance of the same product.
If the user confirms, increase the quantity/count of the product in the cart.
If the user cancels, do nothing.
If the product does not exist, add the product to the cart.