feat: Add 'cart' state in 'LoginContext'.
Add a new state variable called 'cart' using the 'useState' hook in the 'LoginContext' component.
Fetch the count of products in cart using the 'getProductInCartCount' api and set the count in the 'cart' state variable.
Define the 'handleAddCart' function to increament cart count.
Return the 'cart' satate variable and the 'handleAddCart' function from the 'LoginContext' component.