feat: Render products in cart.
Define the 'getProductInCart' function to retrieve the products in the cart using the Firebase API in the firebase.js file.
Import the 'getProductInCart' function from the firebase.js file in the 'Cart' component.
Define a new state variable called 'products' state variable useing the useState hook in the 'Cart' component.
Call the 'getProductsInCart' function and pass it to the 'products' state variable using the 'setProducts' function in the 'Cart' component.
Render the 'products' in the 'Cart' component.