feat: Decrement product in cart.
Define the 'removeCart' function to remove the product in the cart using Firebase API.
When the users clicks the minus button, retrieve the corresponding product information from the cart.
Decrement the count of the product by one using the 'setProduct' function.
If the count of the product is less than one, call the 'removeCart' function from the Firebase API to remove the product information in the cart.
Otherwise, call the 'addCart' function from the Firebase API to update the product information in the cart.
Recalculate the total price by decrementing the product price multiplied with its count using the 'setTotalPrice' function.