refactor: Retreive isAdmin data from user state.
Modify the adminUser function in the 'firebase.js' file.
Invoke the function within the 'onUserStateChange' function when user data exists.
Check if the user's UID is included in the admin list retreived from databse.
If the user is an admin, set the 'isAdmin' property to true in the user state.
Retreive the 'isAdmin' data from user state in the 'Header' and 'NewProduct' component.