refactor: Protect routes from logout users.
Create a new file called 'ProtectedRoute' and define the 'ProtectedRoute' component.
Render its children component or redirect to the '/' path base on the user's authentication status in the 'ProtectedRout' component.
Wrap the 'Cart' and 'NewProdect' component with the 'ProtectedRoute' component to protect these routes
from being accessed by logged out users or non-admin users
Remove the redirect code in the 'NewProduct' component.