feat: Display success notification after product insertion.
Add a new state variable called 'isSuccess' in the 'NewProduct' component and initialize its value as false.
When the product is successfully inserted, set the value of 'isSuccess' to true to display the notification message "The product was inserted successfully".
After 4 seconds, reset the value of 'isSuccess' to false using the 'setTimeout' function.