feat: Render notification of successful cart addtion.
Create a new file called 'Notificaiton.jsx' and define the 'Notification.jsx' component to render the notification title.
Implement the 'Notification' component in the 'Product' page component to display a notification when a product is successfully added to cart.
Implement the 'Notification' component in the 'NewProduct' Page component to display a notification when a product is successfully inserted into database.